requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.0
Pillow==5.4.1
pytz==2018.9
redis==3.1.0
six==1.12.0
uWSGI==2.0.18
vine==1.2.0
Werkzeug==0.14.1
